+++ nihtest/File.py
@@ -3,6 +3,7 @@ import pathlib
 import shutil
 
 from nihtest import Command
+from nihtest import Environment
 from nihtest import Configuration
 from nihtest import Utility
 
@@ -36,7 +37,7 @@ class File:
             if key in configuration.comparators:
                 comparator = configuration.comparators[key]
                 arguments = comparator[1:] + [input_file_name, output_file_name]
-                command = Command.Command(configuration.find_program(comparator[0]), arguments)
+                command = Command.Command(configuration.find_program(comparator[0]), arguments, environment=Environment.Environment(configuration).environment)
                 command.run()
                 if command.exit_code != 0:
                     print(f"{self.name} differs:")