+++ NEWS.md
@@ -1,3 +1,8 @@
-# 1.0 [Unreleased]
+# 1.1.0 [Unreleased] 
+
+- Improve environment variable handling.
+- Reject config file with unknown sections or directives.
+
+# 1.0.0 [2023-06-09]
 
 - Initial public release of Python version of nihtest.